+++ changelog.md
@@ -1,2 +1,41 @@
+# v1.1.6
+
+* Removed Node IDs dependancy (thanks hiimjustin000)
+* Changed button to be the correct texture (thanks GDColon)
+* Added change custom songs location setting support
+
+# v1.1.5
+
+* Fixed crash
+
+# v1.1.4
+
+* Fixed a crash on android when downloading SFX
+* Fixed visual bug with popup text
+* Fixed bug with trash button not working
+
+# v1.1.3
+
+* Fixed a crash caused by deleting the song currently playing
+
+# v1.1.2
+
+* Fixed bug
+
+# v1.1.1
+
+* Added Android and Mac support
+
+# v1.1.0
+
+* Changed name from "Song & SFX Trash Can" to "Song Trash Can"
+* New logo
+* Added popup with options to delete songs and SFX
+* Added mod settings button to popups
+* Added "Auto Select Songs", "Auto Select SFX" and "Hide Settings Button" mod settings
+* Removed "Delete SFX" setting
+* Added mod tags
+* Updated mod description
+
 # 1.0.0
-- Edit this file to change your mod's changelog.
+* Release